improved perf_test to measure steady state behavior improved skiplist random level selection added buitin random number generator; some platforms don't provide a good one improved heuristics for hashtable resize

+++ include/runtime.h
@@ -1,4 +1,4 @@
-/* 
+/*
  * Written by Josh Dybnis and released to the public domain, as explained at
  * http://creativecommons.org/licenses/publicdomain
  */
@@ -11,7 +11,7 @@
 extern DECLARE_THREAD_LOCAL(tid_, int);
 
 int nbd_thread_create (pthread_t *restrict thread, int thread_id, void *(*start_routine)(void *), void *restrict arg);
-int nbd_rand (void);
+uint64_t nbd_rand (void);
 uint64_t nbd_rand_seed (int i);
 int nbd_next_rand (uint64_t *r);
